MacLeish Data Plot 1

Data on change in tree diameter (in centimeters) for parasitic Hemlock Woolly Adelgid dominated areas on the Western side of MacLeish. Tree diameter was measured at 1.4 meters high above the ground.

Usage

tree_diameter1

Format

Module

Module number that represents one of the 10 subplot modules that are 110 m and 20 x 50 m. There are five 10 x 10 modules along central 50 m axis.

Tag

Tag numbers used to identify each tree.

Species

Tree species include Red Maple (Acer rubrum), Sweet Birch (Betula lenta), Paper Birch (Betula papyrifera), American Beech (Fagus grandifolia), American witch-hazel (Hamamelis virginiana), Eastern White Pine (Pinus strobus), Northern Red Oak (Quercus rubra), and Eastern Hemlock (Tsuga canadensis).

Position

Stage of growth for each individual tree. The emergent position are the tallest trees, followed by canopy, subcanopy, and finally, sapling trees, which are the smallest trees.

Year

Data collected from 2010-2015 in the fall semesters.

Notes

Notes collected on the wellbeing of tree species, including notes on death or poor health.

Diameter

Tree diameter measured in centimeters and at 1.4 meters high above the ground.

People

Data obtained from Jesse Bellemare and Smith College students from BIO364-365 courses.

See also